+++ src/Program.fs
@@ -1,5 +1,6 @@
-﻿// Learn more about F# at http://fsharp.org
+// Learn more about F# at http://fsharp.org
 
+open System
 open MF.ConsoleApplication
 open MF.ImageManager
 
@@ -12,46 +13,45 @@ let main argv =
         command "prepare" {
             Description = "Prepares images for sorting"
             Help = None
-            Arguments = [
-                Argument.required "source" "Directory you want to search files."
-                Argument.required "target" "Directory you want to copy files to."
-                Argument.optionalArray "exclude" "Directories you want to exclude from searching." None
-            ]
-            Options = [
-                Option.noValue "force" (Some "f") "If set, target directory will NOT be excluded, and images may be overwritten."
-            ]
+            Arguments = Command.PrepareCommand.arguments
+            Options = Command.PrepareCommand.options
             Initialize = None
             Interact = None
-            Execute = fun (input, output) ->
-                let source = input |> Input.getArgumentValue "source"
-                let target = input |> Input.getArgumentValue "target"
-                let exclude = input |> Input.getArgumentValueAsList "exclude"
+            Execute = Command.PrepareCommand.execute
+        }
 
-                if output.IsVerbose() then
-                    output.Table ["Source"; "Target"; "Exclude"] [[source; target; exclude |> sprintf "%A"]]
+        command "about" {
+            Description = "Displays information about the current project."
+            Help = None
+            Arguments = []
+            Options = []
+            Initialize = None
+            Interact = None
+            Execute = fun (_input, output) ->
+                let ``---`` = [ "------------------"; "----------------------------------------------------------------------------------------------" ]
 
-                output.Section (sprintf "Prepare from %s to %s:" source target)
+                output.Table [ AssemblyVersionInformation.AssemblyProduct ] [
+                    [ "Description"; AssemblyVersionInformation.AssemblyDescription ]
+                    [ "Version"; AssemblyVersionInformation.AssemblyVersion ]
 
-                {
-                    Source = source
-                    Target = target
-                    TargetDirMode =
-                        match input with
-                        | Input.IsSetOption "force" _ -> Override
-                        | _ -> Exclude
-                    Exclude =
-                        match exclude with
-                        | [] -> None
-                        | excludeDirs -> Some excludeDirs
-                }
-                |> Prepare.prepareForSorting output
-                |> function
-                    | Ok message ->
-                        output.Success message
-                        ExitCode.Success
-                    | Error message ->
-                        output.Error message
-                        ExitCode.Error
+                    ``---``
+                    [ "Environment" ]
+                    ``---``
+                    [ ".NET Core"; Environment.Version |> sprintf "%A" ]
+                    [ "Command Line"; Environment.CommandLine ]
+                    [ "Current Directory"; Environment.CurrentDirectory ]
+                    [ "Machine Name"; Environment.MachineName ]
+                    [ "OS Version"; Environment.OSVersion |> sprintf "%A" ]
+                    [ "Processor Count"; Environment.ProcessorCount |> sprintf "%A" ]
+
+                    ``---``
+                    [ "Git" ]
+                    ``---``
+                    [ "Branch"; AssemblyVersionInformation.AssemblyMetadata_gitbranch ]
+                    [ "Commit"; AssemblyVersionInformation.AssemblyMetadata_gitcommit ]
+                ]
+
+                ExitCode.Success
         }
     }
     |> run argv