+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+
+## 0.0.0 -
 - Initial implementation
 - Use paket, FAKE and update to dotnet core 3.0
 - Automatically exclude a target directory
 - Add `force` option
+- Allow multiple sources
+- Allow to exclude images from a file
+- Allow to dry-run the preparation
+- Allow to create a target subdirs by image created year/month
+- Allow config file
+- Update to dotnet core 5.0